refactor: vo 패키지 생성
