+++ src/main/java/kitchenpos/domain/Menu.java
@@ -8,6 +8,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Transient;
+import kitchenpos.domain.vo.Price;
 
 @Entity
 public class Menu {
