+++ src/main/java/kitchenpos/domain/vo/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.vo;
 
 import java.math.BigDecimal;
 import java.util.Objects;
